taxonomy/scripts/requirements.txt
# SPDX-License-Identifier: Apache-2.0

instructlab-schema>=0.4.0
